requirements.txt
Flask
Flask-Cors
flask-sock
google-cloud-storage
google-generativeai
python-dotenv
websockets
asyncio
gunicorn